requirements.txt
flask==2.3.3
cryptography==41.0.3
Werkzeug==2.3.7
MarkupSafe==2.1.3
Jinja2==3.1.2
click==8.1.7
itsdangerous==2.1.2
blinker==1.6.2
pytest==7.4.0
python-dotenv==1.0.0